Use a file resource instead of template for config file
Instead of having a template with a .yml.erb, serialize the attributes to YAML
and write them out as a file resource.

Deletes the `agent_` keys that are not used in the config file, but dumps all
of the other attributes.

Fix a place where there were tabs instead of spaces for indentation.

Signed-off-by: Nathan L Smith <smith@nlsmith.com>